examples/basic/pyproject.toml
[project]
name = "basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "python-magic-bin>=0.4.14",
]
